backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
openai==1.12.0
anthropic==0.18.1
requests==2.31.0
pdfplumber==0.10.3
python-docx==1.1.0
PyMuPDF==1.23.8
pandas==2.2.0
aiofiles==23.2.1
